Save the scan dimensions to /frames
In the HDF5 format, save the scan dimensions to /frames as
a list of two attributes (width, height).

This also changes the convention we currently have to save
dimension attributes as a list of two attributes (width, height)
rather than as separate attributes, such as Nx and Ny.

This is not fully tested and may not fully work.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>